ngrx packages and references removed from UI components
