+++ package.json
@@ -21,10 +21,6 @@
     "@angular/platform-browser": "~11.2.0",
     "@angular/platform-browser-dynamic": "~11.2.0",
     "@angular/router": "~11.2.0",
-    "@ngrx/effects": "~11.0.0",
-    "@ngrx/entity": "~11.0.0",
-    "@ngrx/router-store": "~11.0.0",
-    "@ngrx/store": "~11.0.0",
     "@reduxjs/toolkit": "^1.5.1",
     "normalize.css": "~8.0.0",
     "rxjs": "~6.6.0",
@@ -39,8 +35,6 @@
     "@angular-eslint/template-parser": "~1.2.0",
     "@angular/cli": "~11.2.0",
     "@angular/compiler-cli": "~11.2.0",
-    "@ngrx/schematics": "~11.0.0",
-    "@ngrx/store-devtools": "~11.0.0",
     "@types/jasmine": "~3.6.0",
     "@types/node": "~14.14.0",
     "@typescript-eslint/eslint-plugin": "~4.15.0",
