+++ src/app/app.module.ts
@@ -7,7 +7,6 @@ import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { environment } from '../environments/environment';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { CoreModule } from './core';
 
 @NgModule({
   declarations: [AppComponent],
@@ -17,7 +16,6 @@ import { CoreModule } from './core';
     HttpClientModule,
     MatToolbarModule,
     AppRoutingModule,
-    //CoreModule,
   ],
   providers: [{ provide: 'API_URL', useValue: environment.apiUrl }],
   bootstrap: [AppComponent],
