--- src/app/core/core.module.ts
@@ -1,9 +1,0 @@
-import { NgModule } from '@angular/core';
-
-@NgModule({
-  imports: [
-  ],
-})
-export class CoreModule {
-  constructor() { }
-}
