+++ src/app/core/index.ts
@@ -1,1 +1,2 @@
-export * from './core.module';
+export * from './states'
+export * from './reducers'
