+++ src/app/todo/containers/todo-delete-dialog/todo-delete-dialog.component.html
@@ -16,3 +16,4 @@
     Remove
   </button>
 </mat-dialog-actions>
+
