+++ src/app/todo/selectors/todo.selector.ts
@@ -1,11 +1,16 @@
 import { State, adapter } from '../states';
 import { createSelector } from '@reduxjs/toolkit';
 import { RootState } from '../store';
+import { map } from 'rxjs/operators';
 const { selectAll, selectEntities } = adapter.getSelectors();
 
 /**
  * Selectors
  */
+export const select = <T>(
+  subSelect: (state: RootState) => T
+) => map<RootState, T>(state => subSelect(state))
+
 const selectBaseState = (state: RootState) => state.todo
 
 export const getLoading = createSelector(
