+++ src/app/todo/slices/index.ts
@@ -1,2 +1,1 @@
 export * from './todo.reducer';
-export * from './router.reducer'
