+++ src/app/todo/states/todo.state.ts
@@ -7,35 +7,24 @@ import { createEntityAdapter, EntityState } from '@reduxjs/toolkit';
 export const featureName = 'todo';
 
 /**
- * State
-export type State = EntityState<Todo> & {
-  loading: boolean;
-  selectedId?: string;
-  error?: any;
-}
-
-/**
  * Adapter
-export const adapter = createEntityAdapter<Todo>();
-
-/**
- * Initial state
-export const initialState: State = adapter.getInitialState({
-  loading: false
-});
-
-*/
-
+ */
 export const adapter = createEntityAdapter<Todo>({
   sortComparer: (a, b) => a.title.localeCompare(b.title)
 })
 
+/**
+ * State
+ */
 export type State = EntityState<Todo> & {
   loading: boolean;
   selectedId?: string;
   error?: any;
 }
 
+/**
+ * Initial state
+ */
 export const initialState: State = adapter.getInitialState({
   loading: false
 } as State);
